tests/test_osrs_parser.py
"""Tests for the OSRS Wiki Parser module.

This module contains unit tests for the OSRS Wiki Parser functionality,
including fetching data from the wiki and parsing HTML content.
"""

import pytest
import json
import responses
from bs4 import BeautifulSoup
from unittest.mock import patch, MagicMock

from economy_models.osrs_parser import fetch_wiki_data, OSRSWikiParser, OSRS_WIKI_BASE_URL


class TestFetchWikiData:
    """Test suite for the fetch_wiki_data function."""

    @responses.activate
    def test_successful_fetch_with_infobox(self):
        """Test fetching data when an infobox is present."""
        # Mock response with an infobox
        item_name = "Dragon_scimitar"
        mock_html = """
        <html>
            <table class="infobox">
                <tr><th>Released</th><td>29 March 2005</td></tr>
                <tr><th>Members</th><td>Yes</td></tr>
            </table>
        </html>
        """
        
        responses.add(
            responses.GET, 
            f"{OSRS_WIKI_BASE_URL}{item_name}",
            body=mock_html,
            content_type="text/html",
            status=200
        )
        
        result = fetch_wiki_data("Dragon_scimitar")
        data = json.loads(result)
        
        assert "released" in data
        assert data["released"] == "29 March 2005"
        assert "members" in data
        assert data["members"] == "Yes"

    @responses.activate
    def test_fetch_with_fallback_content(self):
        """Test fetching data when no infobox is present."""
        # Mock response with no infobox but with paragraphs
        item_name = "Some_quest"
        mock_html = """
        <html>
            <title>Some Quest - OSRS Wiki</title>
            <div id="mw-content-text">
                <p>This is a quest in Old School RuneScape.</p>
                <p>It requires level 50 in various skills.</p>
            </div>
        </html>
        """
        
        responses.add(
            responses.GET, 
            f"{OSRS_WIKI_BASE_URL}{item_name}",
            body=mock_html,
            content_type="text/html",
            status=200
        )
        
        result = fetch_wiki_data("Some_quest")
        data = json.loads(result)
        
        assert "content" in data
        assert "This is a quest in Old School RuneScape." in data["content"]
        assert "It requires level 50 in various skills." in data["content"]

    @responses.activate
    def test_http_error_handling(self):
        """Test handling of HTTP errors."""
        item_name = "Nonexistent_item"
        
        responses.add(
            responses.GET, 
            f"{OSRS_WIKI_BASE_URL}{item_name}",
            status=404
        )
        
        result = fetch_wiki_data("Nonexistent_item")
        assert "HTTP 404" in result
        assert "error" in result.lower()

    @responses.activate
    def test_network_error_handling(self):
        """Test handling of network errors."""
        item_name = "Network_error_test"
        
        responses.add(
            responses.GET, 
            f"{OSRS_WIKI_BASE_URL}{item_name}",
            body=responses.ConnectionError("Connection refused")
        )
        
        result = fetch_wiki_data("Network_error_test")
        assert "Network error" in result


class TestOSRSWikiParser:
    """Test suite for the OSRSWikiParser class."""

    def test_parse_item_infobox(self):
        """Test parsing an item infobox."""
        sample_html = """
        <div class="mw-parser-output">
            <table class="infobox-item">
                <tbody>
                    <tr><th>Released</th><td>29 March 2005</td></tr>
                    <tr><th>Members</th><td>Yes</td></tr>
                    <tr><th>Quest item</th><td>No</td></tr>
                </tbody>
            </table>
        </div>
        """
        
        parser = OSRSWikiParser(sample_html)
        data = parser.parse_item_infobox()
        
        assert "released" in data
        assert data["released"] == "29 March 2005"
        assert "members" in data
        assert data["members"] == "Yes"
        assert "quest_item" in data
        assert data["quest_item"] == "No"

    def test_parse_combat_bonuses(self):
        """Test parsing combat bonuses from an infobox."""
        sample_html = """
        <div class="mw-parser-output">
            <table class="infobox-item">
                <tbody>
                    <tr><th class="infobox-subheader">Combat bonuses</th></tr>
                    <tr>
                        <td>
                            <table class="infobox-bonuses">
                                <tbody>
                                    <tr><th class="infobox-subheader">Attack bonuses</th></tr>
                                    <tr><td>stab</td><td>+67</td></tr>
                                    <tr><td>slash</td><td>+66</td></tr>
                                </tbody>
                            </table>
                        </td>
                    </tr>
                </tbody>
            </table>
        </div>
        """
        
        parser = OSRSWikiParser(sample_html)
        data = parser.parse_item_infobox()
        
        assert "combat_bonuses" in data
        # Note: The current implementation might not extract these correctly,
        # this test may need adjustment based on the actual implementation behavior

    def test_parse_equipment_requirements(self):
        """Test parsing equipment requirements."""
        sample_html = """
        <div class="mw-parser-output">
            <table class="infobox-item">
                <tbody>
                    <tr><th>Equipment requirements</th>
                    <td>
                        <ul>
                            <li>60 Attack</li>
                            <li>40 Defence</li>
                        </ul>
                    </td></tr>
                </tbody>
            </table>
        </div>
        """
        
        parser = OSRSWikiParser(sample_html)
        data = parser.parse_item_infobox()
        
        assert "requirements" in data
        # Note: The current implementation might not extract these correctly,
        # this test may need adjustment based on the actual implementation behavior

    def test_parse_dragon_scimitar(self):
        """Test the specialized Dragon Scimitar parser."""
        sample_html = """
        <div class="mw-parser-output">
            <table class="infobox-item">
                <tbody>
                    <tr><th class="infobox-header">Dragon scimitar</th></tr>
                    <tr><th>Released</th><td>29 March 2005</td></tr>
                    <tr><th>Members</th><td>Yes</td></tr>
                </tbody>
            </table>
        </div>
        """
        
        parser = OSRSWikiParser(sample_html)
        data = parser.parse_dragon_scimitar()
        
        assert "item_name" in data
        assert data["item_name"] == "Dragon scimitar"
        assert "released" in data
        assert data["released"] == "29 March 2005"
        assert "members" in data
        assert data["members"] == "Yes"

    def test_save_html_for_debug(self, tmp_path):
        """Test saving HTML content for debugging."""
        html_content = "<html><body>Test content</body></html>"
        debug_file = tmp_path / "debug_file.html"
        
        parser = OSRSWikiParser(html_content)
        with patch('builtins.open') as mock_open:
            parser.save_html_for_debug(html_content, str(debug_file))
            mock_open.assert_called_once_with(str(debug_file), 'w', encoding='utf-8')
