requirements.txt
Flask>=2.0.0
PyYAML>=6.0
requests>=2.25.0
beautifulsoup4>=4.9.0
playwright>=1.20.0
boto3>=1.20.0
sendgrid>=6.9.0
psutil>=5.8.0
Flask-SQLAlchemy>=3.0.0
Flask-SocketIO>=5.3.0
SQLAlchemy>=1.4.0
marshmallow>=3.15.0
marshmallow-sqlalchemy>=0.28.0

# Development dependencies
flake8>=5.0.0
pytest>=7.0.0
pytest-mock>=3.8.0
pytest-cov>=4.0.0
responses>=0.23.0
pytest-flask>=1.2.0
PyGithub>=2.1.0